requirements.txt
requests
kafka-python
pyyaml
neo4j
